Added "--ciphers" and "--fingerprint" options to roscon;
Fixed Util::find() when using multiple numbers in a call, or using three consecutive commas (which would previously be normalized to two instead of one).

Changed the default context adjustments Communicator does when a context is not provided to an encrypted connection: The verify_peer and verify_peer_name settings are set to false if there's no cafile or capath on any level, and set to true otherwise. ADH is still set only if the "ciphers" option is not set, or is set to "DEFAULT". This allows users to use a certificate without providing a context, and do so consistently regardless of PHP version - just set the cafile or capath in either php.ini or with stream_context_set_default(), and you're ready.